+++ src/main/java/com/memora/server/repository/UserSettingsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserSettingsRepository extends JpaRepository<UserSettings, Long> {
+public interface UserSettingsRepository extends JpaRepository<UserSettings, Integer> {
 }
